Added mask for phone field and change data type in Address

+++ PracticeTask2016/PracticeTask2016/Entity/Address.cs
@@ -10,8 +10,8 @@ namespace PracticeTask2016.Entity
     public class Address
     {
         public string street { get; set; }
-        public byte house { get; set; }
-        public byte apartment { get; set; }
+        public ushort house { get; set; }
+        public ushort apartment { get; set; }
 
         public override string ToString()
         {
